thread implementation no longer include mega.h

+++ include/mega/thread/cppthread.h
@@ -42,7 +42,7 @@ public:
     virtual void join();
     virtual ~CppThread();
 
-    static uint64_t currentThreadId();
+    static unsigned long long currentThreadId();
 
 protected:
     std::thread *thread;
