+++ src/thread/posixthread.cpp
@@ -19,8 +19,9 @@
  * program.
  */
 
-#include "mega.h"
 #include "mega/thread/posixthread.h"
+#include "mega/logging.h"
+#include <sys/time.h>
 
 #ifdef USE_PTHREAD
 
@@ -41,12 +42,12 @@ void PosixThread::join()
     pthread_join(*thread, NULL);
 }
 
-uint64_t PosixThread::currentThreadId()
+unsigned long long PosixThread::currentThreadId()
 {
 #if defined(_WIN32) && !defined(__WINPTHREADS_VERSION)
-    return (uint64_t) pthread_self().x;
+    return (unsigned long long) pthread_self().x;
 #else
-    return (uint64_t) pthread_self();
+    return (unsigned long long) pthread_self();
 #endif
 }
